requirements.txt
# package>=version
python-dotenv==0.12.0
PyMySQL==0.9.3
requests==2.23.0
python-dateutil==2.8.1